added tenant date time until

+++ modules/lib-multi-tenant/src/main/kotlin/cn/sunline/saas/multi_tenant/jpa/TenantListener.kt
@@ -2,6 +2,7 @@ package cn.sunline.saas.multi_tenant.jpa
 
 import cn.sunline.saas.global.util.ContextUtil
 import cn.sunline.saas.global.util.getTenant
+import cn.sunline.saas.global.exception.TenantUninitializedException
 import cn.sunline.saas.multi_tenant.model.MultiTenant
 import org.springframework.stereotype.Component
 import javax.persistence.PrePersist
@@ -17,6 +18,6 @@ class TenantListener<T : MultiTenant> {
 
     @PrePersist
     fun prePersist(o: T) {
-        o.setTenantId(ContextUtil.getTenant())
+        o.setTenantId(ContextUtil.getTenant().toLong())
     }
 }
